moved hero section assets
