--- apps/main-landing/src/components/hero-section/assets/index.ts
@@ -1,4 +1,0 @@
-import lines from './lines.png';
-import tokensFlow from './tokens-flow.png';
-
-export { lines, tokensFlow };
