+++ apps/main-landing/src/components/hero-section/components/background-lines/background-lines.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
 
-import { lines } from '../assets';
+import  lines  from './lines.png';
 
 export const BackgroundLines = () => {
   return (
