+++ apps/main-landing/src/components/hero-section/components/background-lines/lines.png
