+++ apps/main-landing/src/components/hero-section/components/tokens-flow/tokens-flow-image.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
 
-import { tokensFlow } from '../assets';
+import tokensFlow from './tokens-flow.png';
 
 export const TokensFlowImage = () => {
   return (
